Network.AWS.CertificateManager.AddTagsToCertificate

Description

Adds one or more tags to an ACM Certificate. Tags are labels that you can use to identify and organize your AWS resources. Each tag consists of a key and an optional value . You specify the certificate on input by its Amazon Resource Name (ARN). You specify the tag by using a key-value pair.

You can apply a tag to just one certificate if you want to identify a specific characteristic of that certificate, or you can apply the same tag to multiple certificates if you want to filter for a common relationship among those certificates. Similarly, you can apply the same tag to multiple resources if you want to specify a relationship among those resources. For example, you can add the same tag to an ACM Certificate and an Elastic Load Balancing load balancer to indicate that they are both used by the same website. For more information, see Tagging ACM Certificates .

To remove one or more tags, use the RemoveTagsFromCertificate action. To view all of the tags that have been applied to the certificate, use the ListTagsForCertificate action.

Creating a Request

addTagsToCertificate

Arguments

:: Text
attcCertificateARN

-> NonEmpty Tag
attcTags

-> AddTagsToCertificate

Creates a value of AddTagsToCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

attcCertificateARN - String that contains the ARN of the ACM Certificate to which the tag is to be applied. This must be of the form: arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012 For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces .

attcTags - The key-value pair that defines the tag. The tag value is optional.

data AddTagsToCertificate

See: addTagsToCertificate smart constructor.

Request Lenses

attcCertificateARN :: Lens' AddTagsToCertificate Text

String that contains the ARN of the ACM Certificate to which the tag is to be applied. This must be of the form: arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012 For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces .

attcTags :: Lens' AddTagsToCertificate (NonEmpty Tag)

The key-value pair that defines the tag. The tag value is optional.

Destructuring the Response

addTagsToCertificateResponse :: AddTagsToCertificateResponse

Creates a value of AddTagsToCertificateResponse with the minimum fields required to make a request.

data AddTagsToCertificateResponse

See: addTagsToCertificateResponse smart constructor.