Network.AWS.CertificateManager.GetCertificate

Description

Retrieves an ACM Certificate and certificate chain for the certificate specified by an ARN. The chain is an ordered list of certificates that contains the root certificate, intermediate certificates of subordinate CAs, and the ACM Certificate. The certificate and certificate chain are base64 encoded. If you want to decode the certificate chain to see the individual certificate fields, you can use OpenSSL.

Creating a Request

getCertificate

Arguments

:: Text
gcCertificateARN

-> GetCertificate

Creates a value of GetCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcCertificateARN - String that contains a certificate ARN in the following format: arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012 For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces .

data GetCertificate

See: getCertificate smart constructor.

Request Lenses

gcCertificateARN :: Lens' GetCertificate Text

String that contains a certificate ARN in the following format: arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012 For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces .

Destructuring the Response

getCertificateResponse

Arguments

:: Int
gcrsResponseStatus

-> GetCertificateResponse

Creates a value of GetCertificateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcrsCertificate - String that contains the ACM Certificate represented by the ARN specified at input.

gcrsCertificateChain - The certificate chain that contains the root certificate issued by the certificate authority (CA).

gcrsResponseStatus - -- | The response status code.

data GetCertificateResponse

See: getCertificateResponse smart constructor.

Response Lenses

gcrsCertificate :: Lens' GetCertificateResponse (Maybe Text)

String that contains the ACM Certificate represented by the ARN specified at input.

gcrsCertificateChain :: Lens' GetCertificateResponse (Maybe Text)

The certificate chain that contains the root certificate issued by the certificate authority (CA).

gcrsResponseStatus :: Lens' GetCertificateResponse Int

- | The response status code.