Network.AWS.CertificateManager.ResendValidationEmail

Description

Resends the email that requests domain ownership validation. The domain owner or an authorized representative must approve the ACM Certificate before it can be issued. The certificate can be approved by clicking a link in the mail to navigate to the Amazon certificate approval website and then clicking I Approve . However, the validation email can be blocked by spam filters. Therefore, if you do not receive the original mail, you can request that the mail be resent within 72 hours of requesting the ACM Certificate. If more than 72 hours have elapsed since your original request or since your last attempt to resend validation mail, you must request a new certificate. For more information about setting up your contact email addresses, see Configure Email for your Domain .

Creating a Request

resendValidationEmail

Arguments

:: Text
rveCertificateARN

-> Text
rveDomain

-> Text
rveValidationDomain

-> ResendValidationEmail

Creates a value of ResendValidationEmail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rveCertificateARN - String that contains the ARN of the requested certificate. The certificate ARN is generated and returned by the RequestCertificate action as soon as the request is made. By default, using this parameter causes email to be sent to all top-level domains you specified in the certificate request. The ARN must be of the form: arn:aws:acm:us-east-1:123456789012:certificate/12345678-1234-1234-1234-123456789012

rveDomain - The fully qualified domain name (FQDN) of the certificate that needs to be validated.

rveValidationDomain - The base validation domain that will act as the suffix of the email addresses that are used to send the emails. This must be the same as the Domain value or a superdomain of the Domain value. For example, if you requested a certificate for site.subdomain.example.com and specify a ValidationDomain of subdomain.example.com , ACM sends email to the domain registrant, technical contact, and administrative contact in WHOIS and the following five addresses: * adminsubdomain.example.com * administratorsubdomain.example.com * hostmastersubdomain.example.com * postmastersubdomain.example.com * webmaster@subdomain.example.com

data ResendValidationEmail

See: resendValidationEmail smart constructor.

Request Lenses

rveCertificateARN :: Lens' ResendValidationEmail Text

String that contains the ARN of the requested certificate. The certificate ARN is generated and returned by the RequestCertificate action as soon as the request is made. By default, using this parameter causes email to be sent to all top-level domains you specified in the certificate request. The ARN must be of the form: arn:aws:acm:us-east-1:123456789012:certificate/12345678-1234-1234-1234-123456789012

rveDomain :: Lens' ResendValidationEmail Text

The fully qualified domain name (FQDN) of the certificate that needs to be validated.

rveValidationDomain :: Lens' ResendValidationEmail Text

The base validation domain that will act as the suffix of the email addresses that are used to send the emails. This must be the same as the Domain value or a superdomain of the Domain value. For example, if you requested a certificate for site.subdomain.example.com and specify a ValidationDomain of subdomain.example.com , ACM sends email to the domain registrant, technical contact, and administrative contact in WHOIS and the following five addresses: * adminsubdomain.example.com * administratorsubdomain.example.com * hostmastersubdomain.example.com * postmastersubdomain.example.com * webmaster@subdomain.example.com

Destructuring the Response

resendValidationEmailResponse :: ResendValidationEmailResponse

Creates a value of ResendValidationEmailResponse with the minimum fields required to make a request.

data ResendValidationEmailResponse

See: resendValidationEmailResponse smart constructor.