Network.AWS.CertificateManager.Types

Description

Service Configuration

certificateManager :: Service

API version 2015-12-08 of the Amazon Certificate Manager SDK configuration.

Errors

_InvalidTagException :: AsError a => Getting (First ServiceError) a ServiceError

One or both of the values that make up the key-value pair is not valid. For example, you cannot specify a tag value that begins with aws: .

_InvalidDomainValidationOptionsException :: AsError a => Getting (First ServiceError) a ServiceError

One or more values in the DomainValidationOption structure is incorrect.

_TooManyTagsException :: AsError a => Getting (First ServiceError) a ServiceError

The request contains too many tags. Try the request again with fewer tags.

_RequestInProgressException :: AsError a => Getting (First ServiceError) a ServiceError

The certificate request is in process and the certificate in your account has not yet been issued.

_InvalidARNException :: AsError a => Getting (First ServiceError) a ServiceError

The requested Amazon Resource Name (ARN) does not refer to an existing resource.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified certificate cannot be found in the caller's account, or the caller's account cannot be found.

_InvalidStateException :: AsError a => Getting (First ServiceError) a ServiceError

Processing has reached an invalid state. For example, this exception can occur if the specified domain is not using email validation, or the current certificate status does not permit the requested operation. See the exception message returned by ACM to determine which state is not valid.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

An ACM limit has been exceeded. For example, you may have input more domains than are allowed or you've requested too many certificates for your account. See the exception message returned by ACM to determine which limit you have violated. For more information about ACM limits, see the Limits topic.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The certificate is in use by another AWS service in the caller's account. Remove the association and try again.

CertificateStatus

data CertificateStatus

Constructors

CSExpired

CSFailed

CSInactive

CSIssued

CSPendingValidation

CSRevoked

CSValidationTimedOut

CertificateType

data CertificateType

Constructors

AmazonIssued

Imported

DomainStatus

data DomainStatus

Constructors

Failed

PendingValidation

Success

FailureReason

data FailureReason

Constructors

AdditionalVerificationRequired

DomainNotAllowed

InvalidPublicDomain

NoAvailableContacts

Other

KeyAlgorithm

data KeyAlgorithm

Constructors

EcPRIME256V1

Rsa1024

Rsa2048

RenewalStatus

data RenewalStatus

Constructors

RSFailed

RSPendingAutoRenewal

RSPendingValidation

RSSuccess

RevocationReason

data RevocationReason

Constructors

AACompromise

AffiliationChanged

CaCompromise

CertificateHold

CessationOfOperation

KeyCompromise

PrivilegeWithdrawn

RemoveFromCrl

Superceded

Unspecified

CertificateDetail

data CertificateDetail

Contains metadata about an ACM certificate. This structure is returned in the response to a DescribeCertificate request.

See: certificateDetail smart constructor.

certificateDetail :: CertificateDetail

Creates a value of CertificateDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdSubject - The name of the entity that is associated with the public key contained in the certificate.

cdStatus - The status of the certificate.

cdFailureReason - The reason the certificate request failed. This value exists only when the certificate status is FAILED . For more information, see Certificate Request Failed in the AWS Certificate Manager User Guide .

cdSubjectAlternativeNames - One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate. The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

cdInUseBy - A list of ARNs for the AWS resources that are using the certificate. A certificate can be used by multiple AWS resources.

cdCreatedAt - The time at which the certificate was requested. This value exists only when the certificate type is AMAZON_ISSUED .

cdCertificateARN - The Amazon Resource Name (ARN) of the certificate. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

cdSerial - The serial number of the certificate.

cdImportedAt - The date and time at which the certificate was imported. This value exists only when the certificate type is IMPORTED .

cdRevokedAt - The time at which the certificate was revoked. This value exists only when the certificate status is REVOKED .

cdNotBefore - The time before which the certificate is not valid.

cdRevocationReason - The reason the certificate was revoked. This value exists only when the certificate status is REVOKED .

cdDomainName - The fully qualified domain name for the certificate, such as www.example.com or example.com.

cdRenewalSummary - Contains information about the status of ACM's managed renewal for the certificate. This field exists only when the certificate type is AMAZON_ISSUED .

cdKeyAlgorithm - The algorithm that was used to generate the key pair (the public and private key).

cdType - The source of the certificate. For certificates provided by ACM, this value is AMAZON_ISSUED . For certificates that you imported with ImportCertificate , this value is IMPORTED . ACM does not provide managed renewal for imported certificates. For more information about the differences between certificates that you import and those that ACM provides, see Importing Certificates in the AWS Certificate Manager User Guide .

cdIssuedAt - The time at which the certificate was issued. This value exists only when the certificate type is AMAZON_ISSUED .

cdSignatureAlgorithm - The algorithm that was used to sign the certificate.

cdDomainValidationOptions - Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED .

cdIssuer - The name of the certificate authority that issued and signed the certificate.

cdNotAfter - The time after which the certificate is not valid.

cdSubject :: Lens' CertificateDetail (Maybe Text)

The name of the entity that is associated with the public key contained in the certificate.

cdStatus :: Lens' CertificateDetail (Maybe CertificateStatus)

The status of the certificate.

cdFailureReason :: Lens' CertificateDetail (Maybe FailureReason)

The reason the certificate request failed. This value exists only when the certificate status is FAILED . For more information, see Certificate Request Failed in the AWS Certificate Manager User Guide .

cdSubjectAlternativeNames :: Lens' CertificateDetail (Maybe (NonEmpty Text))

One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate. The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

cdInUseBy :: Lens' CertificateDetail [Text]

A list of ARNs for the AWS resources that are using the certificate. A certificate can be used by multiple AWS resources.

cdCreatedAt :: Lens' CertificateDetail (Maybe UTCTime)

The time at which the certificate was requested. This value exists only when the certificate type is AMAZON_ISSUED .

cdCertificateARN :: Lens' CertificateDetail (Maybe Text)

The Amazon Resource Name (ARN) of the certificate. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

cdSerial :: Lens' CertificateDetail (Maybe Text)

The serial number of the certificate.

cdImportedAt :: Lens' CertificateDetail (Maybe UTCTime)

The date and time at which the certificate was imported. This value exists only when the certificate type is IMPORTED .

cdRevokedAt :: Lens' CertificateDetail (Maybe UTCTime)

The time at which the certificate was revoked. This value exists only when the certificate status is REVOKED .

cdNotBefore :: Lens' CertificateDetail (Maybe UTCTime)

The time before which the certificate is not valid.

cdRevocationReason :: Lens' CertificateDetail (Maybe RevocationReason)

The reason the certificate was revoked. This value exists only when the certificate status is REVOKED .

cdDomainName :: Lens' CertificateDetail (Maybe Text)

The fully qualified domain name for the certificate, such as www.example.com or example.com.

cdRenewalSummary :: Lens' CertificateDetail (Maybe RenewalSummary)

Contains information about the status of ACM's managed renewal for the certificate. This field exists only when the certificate type is AMAZON_ISSUED .

cdKeyAlgorithm :: Lens' CertificateDetail (Maybe KeyAlgorithm)

The algorithm that was used to generate the key pair (the public and private key).

cdType :: Lens' CertificateDetail (Maybe CertificateType)

The source of the certificate. For certificates provided by ACM, this value is AMAZON_ISSUED . For certificates that you imported with ImportCertificate , this value is IMPORTED . ACM does not provide managed renewal for imported certificates. For more information about the differences between certificates that you import and those that ACM provides, see Importing Certificates in the AWS Certificate Manager User Guide .

cdIssuedAt :: Lens' CertificateDetail (Maybe UTCTime)

The time at which the certificate was issued. This value exists only when the certificate type is AMAZON_ISSUED .

cdSignatureAlgorithm :: Lens' CertificateDetail (Maybe Text)

The algorithm that was used to sign the certificate.

cdDomainValidationOptions :: Lens' CertificateDetail (Maybe (NonEmpty DomainValidation))

Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED .

cdIssuer :: Lens' CertificateDetail (Maybe Text)

The name of the certificate authority that issued and signed the certificate.

cdNotAfter :: Lens' CertificateDetail (Maybe UTCTime)

The time after which the certificate is not valid.

CertificateSummary

data CertificateSummary

This structure is returned in the response object of ListCertificates action.

See: certificateSummary smart constructor.

certificateSummary :: CertificateSummary

Creates a value of CertificateSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csCertificateARN - Amazon Resource Name (ARN) of the certificate. This is of the form: arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012 For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces .

csDomainName - Fully qualified domain name (FQDN), such as www.example.com or example.com, for the certificate.

csCertificateARN :: Lens' CertificateSummary (Maybe Text)

Amazon Resource Name (ARN) of the certificate. This is of the form: arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012 For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces .

csDomainName :: Lens' CertificateSummary (Maybe Text)

Fully qualified domain name (FQDN), such as www.example.com or example.com, for the certificate.

DomainValidation

data DomainValidation

Contains information about the validation of each domain name in the certificate.

See: domainValidation smart constructor.

domainValidation

Arguments

:: Text
dvDomainName

-> DomainValidation

Creates a value of DomainValidation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvValidationEmails - A list of email addresses that ACM used to send domain validation emails.

dvValidationStatus - The validation status of the domain name.

dvValidationDomain - The domain name that ACM used to send domain validation emails.

dvDomainName - A fully qualified domain name (FQDN) in the certificate. For example, www.example.com or example.com .

dvValidationEmails :: Lens' DomainValidation [Text]

A list of email addresses that ACM used to send domain validation emails.

dvValidationStatus :: Lens' DomainValidation (Maybe DomainStatus)

The validation status of the domain name.

dvValidationDomain :: Lens' DomainValidation (Maybe Text)

The domain name that ACM used to send domain validation emails.

dvDomainName :: Lens' DomainValidation Text

A fully qualified domain name (FQDN) in the certificate. For example, www.example.com or example.com .

DomainValidationOption

data DomainValidationOption

Contains information about the domain names that you want ACM to use to send you emails to validate your ownership of the domain.

See: domainValidationOption smart constructor.

domainValidationOption

Arguments

:: Text
dvoDomainName

-> Text
dvoValidationDomain

-> DomainValidationOption

Creates a value of DomainValidationOption with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvoDomainName - A fully qualified domain name (FQDN) in the certificate request.

dvoValidationDomain - The domain name that you want ACM to use to send you validation emails. This domain name is the suffix of the email addresses that you want ACM to use. This must be the same as the DomainName value or a superdomain of the DomainName value. For example, if you request a certificate for testing.example.com , you can specify example.com for this value. In that case, ACM sends domain validation emails to the following five addresses: * adminexample.com * administratorexample.com * hostmasterexample.com * postmasterexample.com * webmaster@example.com

dvoDomainName :: Lens' DomainValidationOption Text

A fully qualified domain name (FQDN) in the certificate request.

dvoValidationDomain :: Lens' DomainValidationOption Text

The domain name that you want ACM to use to send you validation emails. This domain name is the suffix of the email addresses that you want ACM to use. This must be the same as the DomainName value or a superdomain of the DomainName value. For example, if you request a certificate for testing.example.com , you can specify example.com for this value. In that case, ACM sends domain validation emails to the following five addresses: * adminexample.com * administratorexample.com * hostmasterexample.com * postmasterexample.com * webmaster@example.com

RenewalSummary

data RenewalSummary

Contains information about the status of ACM's managed renewal for the certificate. This structure exists only when the certificate type is AMAZON_ISSUED .

See: renewalSummary smart constructor.

renewalSummary

Arguments

:: RenewalStatus
rsRenewalStatus

-> NonEmpty DomainValidation
rsDomainValidationOptions

-> RenewalSummary

Creates a value of RenewalSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsRenewalStatus - The status of ACM's managed renewal of the certificate.

rsDomainValidationOptions - Contains information about the validation of each domain name in the certificate, as it pertains to ACM's managed renewal . This is different from the initial validation that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED .

rsRenewalStatus :: Lens' RenewalSummary RenewalStatus

The status of ACM's managed renewal of the certificate.

rsDomainValidationOptions :: Lens' RenewalSummary (NonEmpty DomainValidation)

Contains information about the validation of each domain name in the certificate, as it pertains to ACM's managed renewal . This is different from the initial validation that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED .

Tag

data Tag

A key-value pair that identifies or specifies metadata about an ACM resource.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagValue - The value of the tag.

tagKey - The key of the tag.

tagValue :: Lens' Tag (Maybe Text)

The value of the tag.

tagKey :: Lens' Tag Text

The key of the tag.